+++ R/sanitize_hypothesis.R
@@ -1,6 +1,6 @@
 sanitize_hypothesis <- function(hypothesis, ...) {
     checkmate::assert(
-        checkmate::check_character(hypothesis, pattern = "[=<>]"),
+        checkmate::check_character(hypothesis, pattern = "=|<=|>="),
         checkmate::check_numeric(hypothesis),
         checkmate::check_formula(hypothesis),
         checkmate::check_matrix(hypothesis),
@@ -19,21 +19,21 @@ sanitize_hypothesis <- function(hypothesis, ...) {
     }
 
     if (isTRUE(checkmate::check_character(hypothesis))) {
-        if (isTRUE(grepl("<", hypothesis))) {
-            hypothesis_direction <- "<"
-        } else if (isTRUE(grepl(">", hypothesis))) {
-            hypothesis_direction <- ">"
+        if (isTRUE(grepl("<=", hypothesis))) {
+            hypothesis_direction <- "<="
+        } else if (isTRUE(grepl(">=", hypothesis))) {
+            hypothesis_direction <- ">="
         }
 
-        if (isTRUE(grepl("^[=<>]", hypothesis))) {
-            hypothesis <- sub("^[=<>]", "", hypothesis)
+        if (isTRUE(grepl("^=|^<=|^>=", hypothesis))) {
+            hypothesis <- sub("^=|^<=|^>=", "", hypothesis)
             hypothesis <- trimws(hypothesis)
             hypothesis <- as.numeric(hypothesis)
         }
     }
 
-    if (isTRUE(checkmate::check_character(hypothesis, pattern = "=|>|<"))) {
-        out <- paste(gsub("=|<|>", "-(", hypothesis), ")")
+    if (isTRUE(checkmate::check_character(hypothesis, pattern = "=|<=|>="))) {
+        out <- paste(gsub("=|<=|>=", "-(", hypothesis), ")")
         attr(out, "label") <- hypothesis
         hypothesis <- out
     } else if (isTRUE(checkmate::check_matrix(hypothesis))) {